+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 2.0.0
+- Added i2, i4, i8, i16, i32 and i64 signed integer datatypes
+- Added u1, u2, u4, u8, u16, u32, and u63 unsigned integer datatypes
+- Updated ix and ux datatypes
+- Added getters for new datatypes
+- Added documentation for all API elements
+
 ## 1.0.0
 - Added getters for Signed Integers
 - Added getters for Unsigned Integers